DeepSpeed checkpoint save/load

+++ sockeye/training.py
@@ -21,7 +21,7 @@ import pickle
 import random
 import shutil
 import time
-from collections import deque
+from collections import deque, OrderedDict
 from dataclasses import dataclass
 from typing import Any, Callable, Dict, List, Optional, Iterable, Tuple, Union, Set
 
@@ -29,11 +29,17 @@ import numpy as np
 import torch
 import torch.distributed
 
+# Optional imports. Import errors are not an issue because these modules are
+# only used when certain trainer settings are activated. We check that these
+# modules can be imported before creating the trainer.
+try:
+    import deepspeed
+    import deepspeed.utils.zero_to_fp32
+except ImportError:
+    pass
 try:
     import apex.amp
 except ImportError:
-    # Not an issue because Apex AMP is only used when the trainer setting is
-    # activated. We check that Apex can be imported before creating the trainer.
     pass
 
 from . import average
@@ -66,6 +72,18 @@ class ModelWithLoss(torch.nn.Module):
         self.model = model
         self.traced_model = None  # type: Optional[torch.jit.ScriptModule]
         self.losses = losses
+
+    def load_state_dict(self, state_dict: 'OrderedDict[str, torch.Tensor]', *args, **kwargs):
+        """
+        Ignore missing/unexpected keys for traced modules when loading state.
+        These keys point to the same parameters as the corresponding non-traced
+        versions that are included in the state dictionary.
+        """
+        incompatible_keys = super().load_state_dict(state_dict, strict=False)
+        missing = [key for key in incompatible_keys.missing_keys if 'traced' not in key]
+        unexpected = [key for key in incompatible_keys.unexpected_keys if 'traced' not in key]
+        utils.check_condition(not missing, f'Missing keys: {missing}')
+        utils.check_condition(not unexpected, f'Extra keys: {unexpected}')
 
     def forward(self, source: torch.Tensor,
                 source_length: torch.Tensor,
@@ -213,7 +231,7 @@ class EarlyStoppingTrainer:
             self._scaler = torch.cuda.amp.GradScaler()
         self.using_apex_amp = using_apex_amp
         self.state = None  # type: Optional[TrainState]
-        self._speedometer = Speedometer(frequency=10, auto_reset=False)
+        self._speedometer = Speedometer(frequency=C.MEASURE_SPEED_EVERY, auto_reset=False)
         self._custom_metrics_logger = custom_metrics_logger
         self._tflogger = TensorboardLogger(logdir=os.path.join(self.config.output_dir, C.TENSORBOARD_NAME))
         self.checkpoint_callback = checkpoint_callback
@@ -319,9 +337,8 @@ class EarlyStoppingTrainer:
         metrics and update the best known parameters accordingly.
         """
         self.state.checkpoint += 1
-        # save parameters and evaluate on validation data
-        if utils.is_primary_worker():
-            self._save_params()
+        # All workers call this method
+        self._save_params()
         train_metrics = [lf.metric for lf in self.loss_functions]
         logger.info("Checkpoint [%d]\tUpdates=%d Epoch=%d Samples=%d Time-cost=%.3f Updates/sec=%.3f",
                     self.state.checkpoint, self.state.updates, self.state.epoch,
@@ -341,7 +358,8 @@ class EarlyStoppingTrainer:
                 self._save_optimizer_state(self.best_optimizer_state_fname)
                 self._save_lr_scheduler(self.best_lr_scheduler_fname)
             self._write_and_log_metrics(train_metrics=train_metrics, val_metrics=val_metrics)
-            self._save_training_state(train_iter)
+        # All workers call this method
+        self._save_training_state(train_iter)
         for metric in train_metrics:
             metric.reset()
         if self.checkpoint_callback:
@@ -649,11 +667,40 @@ class EarlyStoppingTrainer:
         """
         Saves model parameters at current checkpoint and optionally cleans up older parameter files to save disk space.
         """
-        self.sockeye_model.save_parameters(self.current_params_fname)
-        cleanup_params_files(self.config.output_dir, self.config.max_params_files_to_keep, self.state.checkpoint,
-                             self.state.best_checkpoint, self.config.keep_initializations,
-                             self.config.max_params_files_to_cache, self.config.cache_metric,
-                             self.config.cache_strategy)
+        if self.using_deepspeed:
+            # Compatibility: Save regular Sockeye parameter checkpoints when
+            # using DeepSpeed (float32, expected format).
+            training_state_dirname = os.path.join(self.config.output_dir, C.TRAINING_STATE_TEMP_DIRNAME)
+            if utils.is_primary_worker() and not os.path.exists(training_state_dirname):
+                os.mkdir(training_state_dirname)
+            torch.distributed.barrier()
+            # Write a temporary DeepSpeed checkpoint
+            self.model_object.save_checkpoint(training_state_dirname)  # type: ignore
+            if utils.is_primary_worker():
+                # Gather the float32 params on CPU
+                state_dict = deepspeed.utils.zero_to_fp32.get_fp32_state_dict_from_zero_checkpoint(
+                    training_state_dirname)
+                # Strip the first prefix from each param name to match the
+                # SockeyeModel's names:
+                # Ex: 'model.encoder.layers...' -> 'encoder.layers...'
+                state_dict = {name[name.find('.') + 1:]: param for (name, param) in state_dict.items()}
+                # Create a temporary CPU SockeyeModel
+                cpu_model = model.SockeyeModel(self.sockeye_model.config)
+                # Load the float32 params. Use non-strict mode because shared
+                # and constant params are not included in the DeepSpeed-
+                # generated state dict.
+                cpu_model.load_state_dict(state_dict, strict=False)
+                # Save the float32 params to disk
+                cpu_model.save_parameters(self.current_params_fname)
+                shutil.rmtree(training_state_dirname)
+        elif utils.is_primary_worker():
+            # Non-DeepSpeed training: primary worker saves parameters
+            self.sockeye_model.save_parameters(self.current_params_fname)
+        if utils.is_primary_worker():
+            cleanup_params_files(self.config.output_dir, self.config.max_params_files_to_keep, self.state.checkpoint,
+                                 self.state.best_checkpoint, self.config.keep_initializations,
+                                 self.config.max_params_files_to_cache, self.config.cache_metric,
+                                 self.config.cache_strategy)
 
     def _save_optimizer_state(self, fname):
         torch.save(self.optimizer.state_dict(), fname)
@@ -679,41 +726,56 @@ class EarlyStoppingTrainer:
         """
         # Create temporary directory for storing the state of the optimization process
         training_state_dirname = os.path.join(self.config.output_dir, C.TRAINING_STATE_TEMP_DIRNAME)
-        if not os.path.exists(training_state_dirname):
+        if utils.is_primary_worker() and not os.path.exists(training_state_dirname):
             os.mkdir(training_state_dirname)
+        if utils.is_distributed():
+            torch.distributed.barrier()
 
-        # (1) Parameters: link current file
-        params_base_fname = C.PARAMS_NAME % self.state.checkpoint
-        params_file = os.path.join(training_state_dirname, C.TRAINING_STATE_PARAMS_NAME)
-        if os.path.exists(params_file):
-            os.unlink(params_file)
-        os.symlink(os.path.join("..", params_base_fname), params_file)
+        if self.using_deepspeed:
+            # DeepSpeed saves parameters, optimizer state, and learning rate
+            # scheduler in a single checkpoint file. All workers need to call
+            # `save_checkpoint()`.
+            self.model_object.save_checkpoint(os.path.join(training_state_dirname,
+                                                           C.TRAINING_STATE_DEEPSPEED))  # type: ignore
+        elif utils.is_primary_worker():
+            # Otherwise, only the primary worker saves the following.
+            # (1) Parameters: link current file
+            params_base_fname = C.PARAMS_NAME % self.state.checkpoint
+            params_file = os.path.join(training_state_dirname, C.TRAINING_STATE_PARAMS_NAME)
+            if os.path.exists(params_file):
+                os.unlink(params_file)
+            os.symlink(os.path.join("..", params_base_fname), params_file)
 
-        # (2) Optimizer state
-        opt_state_fname = os.path.join(training_state_dirname, C.OPT_STATE_LAST)
-        self._save_optimizer_state(opt_state_fname)
+            # (2) Optimizer state
+            opt_state_fname = os.path.join(training_state_dirname, C.OPT_STATE_LAST)
+            self._save_optimizer_state(opt_state_fname)
 
-        # (3) Data iterator
+            # (3) lr_scheduler
+            lr_scheduler_fname = os.path.join(training_state_dirname, C.LR_SCHEDULER_LAST)
+            self._save_lr_scheduler(lr_scheduler_fname)
+
+        # Secondary workers are done. Only the primary worker runs everything
+        # beyond this point.
+        if not utils.is_primary_worker():
+            return
+
+        # (4) Data iterator
         train_iter.save_state(os.path.join(training_state_dirname, C.BUCKET_ITER_STATE_NAME))
 
-        # (4) Random generators
+        # (5) Random generators
         # RNG states: python, numpy, torch
         with open(os.path.join(training_state_dirname, C.RNG_STATE_NAME), "wb") as fp:
             pickle.dump(random.getstate(), fp)
             pickle.dump(np.random.get_state(), fp)
             pickle.dump(torch.random.get_rng_state(), fp)
 
-        # (5) Training state
+        # (6) Training state
         self.state.save(os.path.join(training_state_dirname, C.TRAINING_STATE_NAME))
 
-        # (5.5) lr_scheduler
-        lr_scheduler_fname = os.path.join(training_state_dirname, C.LR_SCHEDULER_LAST)
-        self._save_lr_scheduler(lr_scheduler_fname)
-
-        # (6) AMP grad scaler state
-        if self.using_amp and not self.using_deepspeed:
+        # (7) AMP grad scaler state
+        if self.using_amp:
             torch.save(self._scaler.state_dict(), os.path.join(training_state_dirname, C.GRAD_SCALER_STATE_NAME))
-        if self.using_apex_amp and not self.using_deepspeed:
+        if self.using_apex_amp:
             torch.save(apex.amp.state_dict(), os.path.join(training_state_dirname, C.APEX_AMP_STATE_NAME))
 
         # First we rename the existing directory to minimize the risk of state
@@ -738,36 +800,42 @@ class EarlyStoppingTrainer:
         Loads the full training state from disk.
         :param train_iter: training data iterator.
         """
-        # (1) Parameters
-        params_fname = os.path.join(self.training_state_dirname, C.TRAINING_STATE_PARAMS_NAME)
-        self.sockeye_model.load_parameters(params_fname, device=self.device, allow_missing=False, ignore_extra=False)
+        if self.using_deepspeed:
+            # DeepSpeed loads parameters, optimizer state, and learning rate
+            # scheduler from a single checkpoint file.
+            _, _ = self.model_object.load_checkpoint(os.path.join(self.training_state_dirname,
+                                                                  C.TRAINING_STATE_DEEPSPEED))  # type: ignore
+        else:
+            # (1) Parameters
+            params_fname = os.path.join(self.training_state_dirname, C.TRAINING_STATE_PARAMS_NAME)
+            self.sockeye_model.load_parameters(params_fname, device=self.device, allow_missing=False, ignore_extra=False)
 
-        # (2) Optimizer states
-        opt_state_fname = os.path.join(self.training_state_dirname, C.OPT_STATE_LAST)
-        self._load_optimizer_state(opt_state_fname)
+            # (2) Optimizer states
+            opt_state_fname = os.path.join(self.training_state_dirname, C.OPT_STATE_LAST)
+            self._load_optimizer_state(opt_state_fname)
 
-        # (3) Data Iterator
+            # (3) lr_scheduler
+            lr_scheduler_fname = os.path.join(self.training_state_dirname, C.LR_SCHEDULER_LAST)
+            self._load_lr_scheduler(lr_scheduler_fname)
+
+        # (4) Data Iterator
         train_iter.load_state(os.path.join(self.training_state_dirname, C.BUCKET_ITER_STATE_NAME))
 
-        # (4) Random generators
+        # (5) Random generators
         # RNG states: python, numpy, torch
         with open(os.path.join(self.training_state_dirname, C.RNG_STATE_NAME), "rb") as fp:
             random.setstate(pickle.load(fp))
             np.random.set_state(pickle.load(fp))
             torch.random.set_rng_state(pickle.load(fp))
 
-        # (5) Training state
+        # (6) Training state
         self.state = TrainState.load(os.path.join(self.training_state_dirname, C.TRAINING_STATE_NAME))
 
-        # (5.5) lr_scheduler
-        lr_scheduler_fname = os.path.join(self.training_state_dirname, C.LR_SCHEDULER_LAST)
-        self._load_lr_scheduler(lr_scheduler_fname)
-
-        # (6) AMP grad scaler state
-        if self.using_amp and not self.using_deepspeed:
+        # (7) AMP grad scaler state
+        if self.using_amp:
             self._scaler.load_state_dict(
                 torch.load(os.path.join(self.training_state_dirname, C.GRAD_SCALER_STATE_NAME)))
-        if self.using_apex_amp and not self.using_deepspeed:
+        if self.using_apex_amp:
             apex.amp.load_state_dict(torch.load(os.path.join(self.training_state_dirname, C.APEX_AMP_STATE_NAME)))
 
         logger.info("Training State: epoch=%d, checkpoint=%d batches=%d updates=%d best_metric=%.2f, " \
